qt: (hotkeys) add longtext tooltip to first column of each row
this makes longtext of hotkey actions available to users, helping them to
understand the purpose of the actions listed in the table.

this does not interfere at all with the tooltip giving instructions on
how to perform editing, that still appears when hovering over the other
two columns. no tooltip existed for the first column previously.