qt: (hotkeys) clarify table label
let's start the text with something describing the actual purpose
of the overall control, rather than leaping directly into
instructions on how to use it. this may not be so important for
where the table is shown in the simple view, but makes a notable
difference to use in the advanced view, where it looks very odd
without it.

also, let's clarify those instructions a bit.