qt: (hotkeys) use comma+space as visible multi-binding separator
keep the existing separator (tab) for stored values, but swap with
comma+space separator for display. it works much better this way imo and
is much less confusing. it also happens to match the style used by the
`QKeySequenceEdit` control.

we thus avoid the problems that the default multi-key bindings create, as
is the default on Linux currently, whereby with play/pause for example you
see "Space       Media Play Pause" which obviously looks odd. now you will
see "Space, Media Play Pause".

if users assign the actual comma key in a multi-key binding, it should
not really be a problem for them to grasp what's what. for instance
",, Media Play Pause".