core: ensure all hotkey options have longtext
with an earlier commit in this set having added a tooltip using longtext
to the first column of the hotkey editing table, allowing users to get
an additional description to help them understand the purpose of an
action, it would be great for best UX if we have such a tooltip for
each and every entry in the table. of the 113 entries, 9 were missing
longtext and thus missing the tooltip.

it was for this reason i skipped hotkeys in
d75459a35b98cb8e6a5978e776335bd5b87ea52b.

the "Select the hotkey to use..." form was used with that being fairly
common for hotkey longtext.

i used a single common longtext for the four zoom presets, as done
already with bookmarks and subtitle text scaling, reducing translator
burden.