qt: (hotkeys) clear conflict flag upon fresh conflict check
if the edit dialog detects a conflict, the 'conflicts' attribute is set to
true and the dialog updated to warn the user and give them a choice as to
what to do about it.

the point of the 'conflicts' attribute relates to one of the choices given
to the user, which allows forcing the assignment. in that case, this flag
lets `KeySelectorControl::selectKey()` know that it must take action to
remove the assignment from existing keys after the dialog closes, before
storing the new assignment.

another choice available to the user is that of simply entering a different
key/combination. if the user does so, this will trigger another conflict
check. this new conflict check could obviously again find a conflict, in
which case the dialog is updated to reflect the new conflict info. if it
does not find a conflict for this new assignment, it is accepted.

in this latter situation, of an initial conflict, which the user resolves
by making a different choice which then does not conflict, the 'conflicts'
flag was left set to true from the original conflict, which meant that the
above mentioned function would waste time trying to remove the existing
assignment before storing the change.

this fixes that inefficiency by simply resetting the flag at the start of
performing a conflict check.