upnp: fix bad option default
the lowercase "auto" default option value did not match the uppercase
"Auto" in the choice list.

option values should typically be lowercase. this fixes the mistake with
the case of the choice value, whilst maintaining backwards compatibility
with existing CLI usage and from saved settings files. this was the
clear preference for fixing this resulting from the MR discussion.