contrib: Rework meson crossfile generation
Previously the crossfile had a lot of hardcoded flags, instead of the
correct CFLAGS/CXXFLAGS, etc.
This replaces the generation in the Makefile with a simple Python script
instead, which should be fine, given that meson anyway needs Python 3
and that the crossfile is only generated when needed.

(cherry picked from commit 48f6d51f6da3e02754bc1b52e45bd9652d1051a4)
Signed-off-by: Marvin Scholz <epirat07@gmail.com>