player: get vout and vout_order from VIDEO/SPU es_ids
vlc_player_GetVoutFromEsId() can now return the vout used by an SPU es_id.
It also returns the vout order.

The on_vout_changed callback is also used for SPU es_ids. Users could check the
category of the es_id to know if the vout is attached to a VIDEO es or an SPU
one.