opengl: remove persistent support
Not compatible with future push model.

It depended way too much on GPU and drivers implementation and there was no
real proof that it was faster. Buffers were mapped on the CPU side, therefore
an upload to the GPU had to be done in all case. Mapping on the GPU was too
slow because decoders need to access reference frames.