direct3d9: no need for a HWND at all to create & use a device
When presenting the data we use the proper HWND to Present(). When rendering
externally it's the host that needs to take care of that, and we render in a
texture anyway.

The decoder/filters don't seem to need the HWND at all (tested on Win7 and 10).

And simplify the calls to D3D9_CreateExternal().