input: simplify private resource handling
Release() (instead of Terminate()) from the closing thread. Don't hold twice
the ref count for private resources.

This will not change anything when the resource is provided to the
input_thread_t. The Release() call will just decrement the ref count.

This will also not change anything when the input_thread_t create its own
resource. The Release() will actually terminate the resource like before, and
from the closing thread.