filters: set the input/output video context on video filters
The input context pushed by the decoder (and soon by the preceding filter in chain).

The output context is similar to the output video video context set by the decoder.

The filter should hold a reference to the output video context and release it
when it doesn't need it anymore.