input: always call es_out functions from input thread
Locking the es_out can be blocking since it can wait for the decoder.
Therefore, we must always go through an input control to call es_out functions.

This is a partial revert of 1f51e1d214771f7d609d518178c0e4c173dff8dd

When this partially reverted commit said:

"Maybe, it would be better to completely hide all input controls and let the
input implementation decide if a control need to be handled from the MainLoop
thread or can be handled directly."

=> This can only be done when all es_out control are non blocking (maybe with
the 5.0 input buffering rewrite ?)