coreaudio: fix first time_get return values
Don't return a valid delay of 0 when the delay is still not known.

This triggered VLC drift correction from aout_DecSynchronize() and could caused
flushing of first buffers or ressampler being inserted.

This also fixes playback with audio devices having a long delay (>= 350ms).
Depending on few options (like the rate), VLC was never able to catch up with
this first valid delay (so it was flushing in loop).

This commit apply for both macOS and iOS.

cf. https://code.videolan.org/videolan/vlc-ios/issues/344