sdp: fix null format string passed to vsnprintf
When rtcp-mux is used, sdp_AddAttribute is called with a null
format string, which means to write an SDP property attribute
(without value), which is a different from an empty format string,
that writes an SDP value attribute with an empty value.

This fixes vsdp_AddAttribute so that it correctly writes a
property attribute to the SDP description rather than incorrectly
calling vsnprintf with a null format string (an undefined behaviour
which segfaults on Windows) and writing a value attribute with
an empty value.

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>