qt: open_panels: prevent the dialog customization from crashing
Check that the widgets extracted from the Qt widgets tree are valid
before using them, otherwise operator[] would crash on access.

More work would be needed to use a safer and more sustainable way to
customize the dialog box. In the mean time, this prevent the UI from
crashing with the new UI.

It was reported on platform like Suze Leap 15.1 when opening a disk.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>