core: libvlc: Let the medialibrary outlive the playlist
As we will save the playback states in the media library from the
player, we need to ensure the media library is still available when
doing so.
Since the last save can occur while destroying the player and it's
associated input, we need the media library to be destroyed after the
player, and therefor the playlist.