filter_chain: no need to reset the fmt_out each time a filter is added
The last output format is always the format of the last filter. It doesn't
matter if it was allowed to change the format or not. It's either the filter
added or the "chain" filter added to compensate for the format change (should
be adjusted in push mode)

This is how filter_chain_GetFmtOut() is documented "last output format of the
last element in the filter chain" and how it's used.

If there's no filter yet, the output format is the one used to initialize the
filter chain (via filter_chain_Reset).