decoder: factorize the output YUV subsampling constraints fixup
This way other decoder owner (thumbnail, transcode, etc) benefit from the same
fix in a consistent manner.

Now we clean the value directly in the dec->fmt_out whereas before it was
cleaned at the vout level and the decoder didn't know about it.

It should be done before the decoder allocates it's picture pool. This is the
case as the default pool is created during the format_update, which called in
the same function later.

Other decoder owners are not impacted as they don't provide a default pool and
just rely on a direct picture_NewFromFormat() call.