qt: use the new player timer
Qt setup a vlc_player_timer with a minimal period of 500ms. To propagate these
point to the UI, 2 QTimer are used.

 - A time QTimer: it will ask the player when the next second of the media will
   be reached in order to wait accordinly and update the time at the right
   time.

 - A position QTimer: it will update regularly the position, depending on the
   length and rate of the media. Maybe it should also do it according to the
   slider width too.

Before that, the input time was used, it could be delayed by 300ms to 2seconds
depending on your output buffer size (generally, led by audio module buffer
size). Furthermore, the input time was only updated every 250ms. This could
lead to a delay of (250 - 1) ms before updating the right second (and not
taking the output delay into account).