smb2: fix teardown when interrupted
When the input is stopped by the user, there are lot of chances that this
access is waiting for the read command completion (smb2_read_async()).

If the read command is not fully processed (since interrupted by VLC), any
future commands will fail, like smb2_close_async() when the access is finally
closed.

To fix this issue, we switch back to the posix poll() when interrupted to let a
chance to finish the current command in order to be able to close the smb2
session nicely.