prepaser: change options handling
Reminder of the current behavior: the meta fetcher is either always triggered
after the preparser is finished or manually.

This commit aims to give more controls so that the meta fetcher is not
necessarily triggered after a preparsng.

The current behavior of macOS/playlist/mediatree is unchanged (the components
touching the preparser).

Here are the API changes for

libvlc_MetadataRequest():

 - Can't be called without a valid META_REQUEST_OPTION_SCOPE_* flag

 - The flag META_REQUEST_OPTION_SCOPE_* is not fetching meta anymore. Use
   (META_REQUEST_OPTION_SCOPE_*|META_REQUEST_OPTION_FETCH_*) to fetch meta when
   the preparsing ends.

libvlc_ArtRequest():

 - Can't be called without a valid META_REQUEST_OPTION_FETCH_* flag.

 - The META_REQUEST_OPTION_FETCH_NETWORK flag will now only fetch meta via
   network.

libvlc_media_parse_with_options():

 - The flag libvlc_media_parse_* is not fetching meta anymore. Use
   libvlc_media_fetch_* to fetch meta when the preparsing ends.
 - The libvlc_media_fetch_network flag will now only fetch meta via network.