player: rework aout/vout callbacks
Document that vlc_player_t functions should not be called from these callbacks.
A player function could trigger an action on the vout/aout and cause a callback
to be called => deadlock.

Remove the vlc_player_t * argument, it was never used.

Add the audio_output_t *argument for aout callbacks. It's not currently used,
it is added as symmetry with vout callbacks.