dxva: balance the surface API and avoid dirty casts
d3d11va/dxva2 only get a va_surface from the API and deal with the references.

They need to create a copy of the internal picture context so it's not released
when the picture is finished in the display. A copy also increments the surface
refcount, so we need to release one after the copy.