video context: make the video context structure private
The video context will be passed to various modules up to the display where it
will be ultimately released when the last associated picture will be released.
It needs to be refcounted and we need a destructor to do this final release.

By default the video context holds a reference to the matching decoder device.
This reference is released with the video context.