Qt: fix improper dataChanged() signal emission
At https://doc.qt.io/qt-5/qabstractitemmodel.html#dataChanged, it reads:
"This signal is emitted whenever the data in an existing item changes."

My understanding is that emitting dataChanged() is limited to the update of
an existing item. Neither insertion nor deletion of an item have to emit
this signal.

Furthermore, in the case of deletion, this may lead to the following error:

ASSERT failure in QList<T>::operator[]: "index out of range", file /usr/include/x86_64-linux-gnu/qt5/QtCore/qlist.h, line 549

Signed-off-by: Thomas Guillem <thomas@gllm.fr>