macosx: widen stored playback position for longer videos
Existing logic of storing playback position for macOS excludes positions within first/last 5%.

This heuristic is fine for short videos, but for a longer one, say a 6-hour one, it means the position is not stored for the first/last 18 minutes, which is a nuisance.

This change adds an upper limit for the size of excluded positions.

Signed-off-by: Mikhail Gusarov <dottedmag@dottedmag.net>
Signed-off-by: Felix Paul Kühne <felix@feepk.net>