wasapi: start deferred
The idea is to act like PulseAudio and coreaudio recently.

Instead of returning a valid date from the first TimeGet() to let the core
handle the initial delay, the initial delay is handled inside this module by
deferring the Start.

The big advantage of this solution is that this module can now handle a change
of play date.

A Windows Timer is used to defer this start. Audio samples are written while
the AudioClient is not started (behavior handled according to the
documentation). If the audio buffer is full before the AudioClient is started,
the Play function will wait for it.