packetizer: detect AC-3 core + EAC3 extension
This will allow audio output to first try EAC3 and/or fallback to AC3.

cf. https://en.wikipedia.org/wiki/Dolby_Digital_Plus#HD_DVD_and_Blu-ray_Disc

"On Blu-ray Disc, DD+ is an optional codec, and is deployed as an extension to a
"core" AC-3 5.1 audiotrack. The AC-3 core is encoded at 640 kbit/s, carries 5
primary channels (and 1 LFE), and is independently playable as a movie audio
track by any Blu-ray Disc player. The DD+ extension bitstream is used on
players that support it by replacing the rear channels in the 5.1 setup with
higher fidelity versions, along with providing a possible channel extension to
6.1 or 7.1. The complete audio track is allowed a combined bitrate of 1.7
Mbit/s: 640 kbit/s for the AC-3 5.1 core, and 1 Mbit/s for the DD+ extension.
During playback, both the core and extension bitstreams contribute to the final
audio-output, according to rules embedded in the bitstream metadata."