contrib: modplug: Fix endianness check
The autoconf endianness check does not compile with a C++ compiler
due to bogus casts, therefore it is needed to use the C compiler to
check for the host endianness.

This fixes the libmodplug contrib build for Android.

(cherry picked from commit 903c250978153ef8d95a6005def39ab0ad4078f9)
Signed-off-by: David Fuhrmann <dfuhrmann@videolan.org>