video_output: do the format cleaning in vout_Request and vout_GetDevice
vout_GetDevice doesn't need to check the DPB, it only needs to get a decoder
device from a usable window. Only the display needs to DPB size (for now, since
it creates the decoder pool from it).

Only stop the current display if a new one is going to be created (otherwise
the current one is fine or there is none because there's no window).

In vout_GetDevice we don't care about the current display. So we don't need to
use sys->original, we use a locally cleaned video format instead to match what will
be used in the following call to vout_Request.

Try to enable the window before checking if the original format matches. If the
window is not enabled, the sys->original doesn't making sense anyway.