jack: use var_LocationParse instead of custom parser
This patch leverage previously defined VLC variables to parse the MRL
through var_LocationParse instead of custom parsing code. The behaviour
is unchanged when it comes to the previous case of MRL, but it is now
possible to define these variables in VLC settings. However, unknown
variables are not reported anymore as long as var_LocationParse doesn't
notify this.