jack: add new ports and channels safe variables
Those variables are the one from the MRL given to jack access. It adds
documentation for these variables and allow them to be defined outside
of the modules. It will also allow to correctly implement the
enumeration callbacks for the ports variables, to ease the user
experience when selecting the input.