interrupt: restore interrupted to the killed state
If a module is interrupted while waiting in a vlc_*_i11e*() function, the i11e
function will return an error after restoring the interrupted state (from
vlc_interrupt_finish()). This error will be carefully handled by the module
that will close itself. If a second module is in the probe list, any call to a
vlc_*_i11e*() function will act as not interrupted. This cause any following
modules in the probe list to ignore this interrupted state.

To fix this issue, vlc_interrupt_finish() will now restore the interrupted
state to the killed state.