avcodec: vaapi: use the va_surface API to allocate/get surfaces
The va_pool is refcounted. It releases its resources when it's not used anymore.
Either when the last surface from the pool is not used anymore or when it's
closed and no picture were used anymore.

The va module manages its own picture context structure that includes the
common structure used by all vaapi modules (vaapi_pic_context) and the
va_surface that is refcounted (so it can release the pool when the last surface
is used and the close is pending).

The table of VASurfaceID is allocated locally, in the same way it's done in
vlc_vaapi_PoolNew().

pictures are coming from the default allocator, there is no picture_sys_t
attached. But it's not used by vlc_vaapi_PicGetSurface() or
vlc_vaapi_PicGetDisplay(). It's only used by vlc_vaapi_PicAttachContext()
which we don't use anymore on pictures we send.

This is the same API used by the dxva modules.