contrib: glew: drop debug postfix from the library name
With 4d57652c407d4a75e0b36f4e1fbc438def3b5c2e the Cmake BUILD_TYPE is
set to Debug when optimizations are disabled, which is then used by glew
to determine the resulting library name.

This in turn confuses our build of projectM, which uses glew32 as a name
to use when linking.