prefetch: fix wrong assert
When the prefetch buffer is filled, when the buffer offset is 0 and when a seek
is done past the buffer length, history can be bigger than buffer_length.

This could lead to a read corruption but not a buffer overflow since the size
of the circular buffer is not impacted by this bug.