mosaic_bridge: don't attempt to create the filter chain early
Either video_format_update had been called and it missed the filter creation
and didn't update it, so we create the chain before the decoder. Or it was
called after and this initialization was useless and reset afterwards.

Now the filter chain will always use a correct video format and don't try to
create a bogus one when the decoder is opening.