picture: add the video context in the picture context
The creator of the picture_context_t must set the video context if needed.
If set, the picture context must hold a reference to the video context.

When "copying" a picture (using the same surface in another picture context) a
new reference has to be acquired.

Add an inline function to get the video context from a picture.