mediacodec: use the new push model
pictures, context and video context are now allocated from the mediacodec
module. This simplify a lot of things since MediaCodec is naturally a push API.

The mediacodec decoder module, via the decoder device will now decide of the vd
plugin used: android-display or gles2. It depends on how the LibVLC user setup
its surfaces or if the format need an OpenGL transformation (360 videos). The
decoder module need to know this informations before configuring MediaCodec in
order to choose the good Surface. The Surface will be either the SurfaceView
directly (AWindow_Video) or a SurfaceTexture that will be attached to a GL
texture (AWindow_SurfaceTexture) by the android gl converter.

Signed-off-by: Steve Lhomme <robux4@ycbcr.xyz>