macosx: Add -fno-stack-check to workaround Xcode 11 issue
Compiling with Xcode 11 automatically enables the -fstack-check feature
which causes stack checking error due to unaligned stack which seem
to not be true. Lots of projects are affected by this currently
so hopefully this will be fixed in an Xcode update.

(cherry picked from commit 856ddd8d1af955712c7effb1876ca9544b688cab)
Signed-off-by: Marvin Scholz <epirat07@gmail.com>