Qt(menu): fix CheckablelistMenu
Using an actionGroup is wrong, because the checkableListMenu just needs to
comply with the model. For some models (Crop, Zoom, Aspect, ..), it is exclusive (one and only one check), for others (audio, video, subtitle), it may not be exclusive(zero check or multiple checks allowed). But, anyway, the models manage that accordingly and there is not need to also want to manage it locally. Furthermore, for audio, video, subtitle, this breaks everything