vdpau/chroma: store the incoming picture in the history
Rather than a copy of the field with a reference held on the surface of the
actual picture. Instead of releasing the picture after we do a shallow copy of
the surface we release it when the surface/picture is not needed in the history
anymore. It has the same effect as releasing the reference of the field when we
don't need it in the history.

No need to store the forced flag or the date of the picture, we have the actual
picture.