macOS packaging: Add target to package an sdk version for macOS
This adds a separate target to add an SDK version of the macOS
binaries. This package is the result of a destroot'ed 'make install'

It shall serve two purposes:

1) Out of tree modules compilation: Module authors can use the included
pkgconfig files to link against libvlc(core) for their own modules.
This is similar to the existing SDK we have on Windows.

2) Packaging libvlc for our own app: If someone does not want to use
VLCKit, but links against libvlc directly, these precompiled artefacts
can be used for packaging the libs into the own app. Packaging shall
be done in the same way as with VLC.app (i.e.: lib/* and the vlc plugins
directory goes to Contents/Frameworks, share/ goes to
Contents/Resources/share/ if needed). In this case, all ressources are
found by libvlc automatically.

Users can also execute vlc directly from the extracted tarball, after
specifying LD_LIBRARY_PATH and VLC_PLUGIN_PATH. Using the macOS UI
module is not supported.