opengl: converter_sw: reset UNPACK_ROW_LENGTH after usage
Some drivers seem to use this value for glTexSubImage2D even with
GL_PIXEL_UNPACK_BUFFER set to 0. As the rest of the code is not
UNPACK_ROW_LENGTH aware, it is safer to reset it to a non-active value.

For example, this would cause glitches if pictures were uploaded with
this extension but SPU were not, or more generally if another client of
the OpenGL context is updating some textures by itself outside of the
OpenGL renderer.

Looks really close to an OpenGL version of #22267.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>
(cherry picked from commit 045f3ef28f86be7eed530f7079ad5ed182da9144)
Signed-off-by: Thomas Guillem <thomas@gllm.fr>