contrib: lame: disable asserts
As the lame INSTALL says:

"For production use, be sure to compile a "Release" target, with the "maximum
speed" compile option, and #define NDEBUG."

It will fix an assert on psymodel.c:576 that could happen when feeding data
after a flush. There is no possible memory corruption but maybe a read of
invalid data that could trigger an audio glitch. Due to the rarity of this bug
(happening only via chromecast, with some input files and with a lot of seek
requests), I think that such fix is enough.

PS: the contrib WITH_OPTIMIZATION option should be split into OPTIMS and DEBUG.