egl: initialize context for error handling
Otherwise, we end up calling eglDestroyContext with a bad value in the
case the surface cannot be create. Not being able to create the surface
isn't even rare as it can happen if the format we required is not supported.

Harmless regression from f5a766e3699f19f3a37f6331e62bdc5de0e7fd04

Signed-off-by: Thomas Guillem <thomas@gllm.fr>