ios: uiview: add new vout_window module
Refactor code from the ios vout display to create a dedicated UIView
which handle the touch events and window resizing.

The resizing events are emitted after any children UIView layer is added
to the UIView vout window so that it doesn't block the main thread under
display lock while the display module is being created. Indeed, it would
result in a lock inversion and thus potential deadlocks.