medialib: Widen stored playback position for long videos
Existing logic of storing playback position excludes positions within last 5%.

This heuristic is fine for short videos, but for a longer, say 6-hour hones, it
means that the position is not stored for the last 18 minutes, which is a
nuisance.

This change adds an upper limit for the size of excluded positions and makes it
symmetric, so that a video that was stopped just after the start will be
restarted from the beginning.

This is a continuation of a macOS-specific patch that landed as
e3e32ab0119edfd762718ecd9a084aeaef873e2f in branch 3.0

Signed-off-by: Mikhail Gusarov <dottedmag@dottedmag.net>
Signed-off-by: Hugo Beauzée-Luyssen <hugo@beauzee.fr>