avcodec: don't detect texture sizes as needing codec changes
lavc forces the coded size at each update, even though that's not the actual
texture size.

Since 9c5c6d212fc60e887579357f86d50537d512c2d3 we can set the output size of
the module to the texture size so it is handled properly downstream and not
hidden in the video context/picture_sys_t. It doesn't work well when lavc
request a format update when the texture dimensions didn't actually changed.

So we keep a copy of the video format dimensions that lavc thinks we are using
to check for actual changes in what it is feeding us.