macosx: fs panel: Ensure fs panel always stays inside the vout rect
In very rare cases, it was possible to move the fs panel to another
monitor, and it was not recentered once fullscreen was started
on a different monitor again. This led to the fullscreen window
and the fs panel being shown on different screens.

To trigger the bug, the user needed to move a native fullscreen window
to another monitor using Expose, which was undetected by VLC and thus
invalidated the internal state.

This commit adds a second check to detect this case and also
ensure recentering of the panel here.