macosx: Refine animations to show and hide the fullscreen controller
Previously, the fs panel was fading in with a 400ms animation, and
at the same time the 1s timer to hide the panel has started.
This led to a percieved visibility of the panel for only 600 ms approx,
which felt a bit too short as a default value.

Now only start the timer once the animation is fully complete, and
avoid repeated abortions of this animation.